backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
httpx>=0.25.2
python-dotenv>=1.0.0
redis>=5.0.1
pydantic>=2.8.0
python-multipart>=0.0.6